# api/openapi.yaml
openapi: 3.0.3
info:
  title: LatentTone Product API
  description: |
    Phase 2+3 HTTP product surface for LatentTone (auth, sessions, stream/HLS,
    feedback, neighbor playlists, Phase 3C user playlists, embed status).

    **Try it out (Swagger / scripts):** call `POST /api/v1/auth/login` (or register),
    copy the response `token`, click **Authorize**, and paste it as a Bearer token.
    The product SPA uses the HttpOnly cookie `lt_session` only (same-origin; no
    Bearer in JavaScript). Bearer remains supported for non-browser clients.

    SPA and API share one origin (`/app/` + `/api/v1/`) — no CORS. Do not add
    `Access-Control-Allow-Origin: *`.

    Swagger UI is flag-gated (`enable_api_docs`) and is **not** a substitute for
    Phase 3 stream-smoke Gates A–D or `/dev/stream`.
  version: 0.3.1
  contact:
    name: martinsah
  license:
    name: GPL-3.0-only
    url: https://www.gnu.org/licenses/gpl-3.0.html
servers:
  - url: /
    description: Same origin as latenttone serve
  - url: https://latent.lt.lkeng.org
    description: Example reverse-proxied public origin (override via PUBLIC_BASE_URL)
...
paths:
...
  /api/v1/playlists:
    post:
      tags: [playlists]
      summary: Create neighbor playlist
      description: |
        Operator API (auth not required per ADR-005 / ADR-008). Generates a k-NN
        neighbor playlist from `seed_track_id`. When a session cookie/Bearer is
        present, `user_id` is stored on the playlist row. There is no list-all
        GET on this path; fetch by id via `GET /api/v1/playlists/{id}`.
      operationId: createPlaylist
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePlaylistRequest'
            example:
              seed_track_id: 418
              length: 20
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Playlist'
        '400':
          $ref: '#/components/responses/Error'
  /api/v1/playlists/{id}:
    get:
      tags: [playlists]
      summary: Get playlist by id
      description: Operator API (auth not required). Returns playlist metadata and track rows.
      operationId: getPlaylist
      security: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Playlist'
        '404':
          $ref: '#/components/responses/Error'
...
  /api/v1/me/playlists:
    get:
      tags: [me-playlists]
      summary: List my user playlists
      description: Returns `kind=user` playlists owned by the caller (headers only).
      operationId: listMyPlaylists
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MyPlaylistList'
        '401':
          $ref: '#/components/responses/Error'
    post:
      tags: [me-playlists]
      summary: Create empty user playlist
      operationId: createMyPlaylist
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUserPlaylistRequest'
            example:
              name: Late night
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Playlist'
        '400':
          $ref: '#/components/responses/Error'
        '401':
          $ref: '#/components/responses/Error'
...
  /api/v1/me/playlists/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
    get:
      tags: [me-playlists]
      summary: Get my playlist
      description: Owner-only; non-owner receives 404.
      operationId: getMyPlaylist
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Playlist'
        '401':
          $ref: '#/components/responses/Error'
        '404':
          $ref: '#/components/responses/Error'
    patch:
      tags: [me-playlists]
      summary: Rename my playlist
      operationId: renameMyPlaylist
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUserPlaylistRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Playlist'
        '400':
          $ref: '#/components/responses/Error'
        '401':
          $ref: '#/components/responses/Error'
        '404':
          $ref: '#/components/responses/Error'
    delete:
      tags: [me-playlists]
      summary: Delete my playlist
      operationId: deleteMyPlaylist
      responses:
        '200':
          description: Deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
        '401':
          $ref: '#/components/responses/Error'
        '404':
          $ref: '#/components/responses/Error'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: opaque
      description: |
        Opaque session id from login/register JSON field `token`.
        Preferred for Swagger Try it out: Authorize → Bearer `<token>`.
    cookieAuth:
      type: apiKey
      in: cookie
      name: lt_session
      description: |
        HTTP-only cookie set on register/login. Same-origin browser clients use
        this automatically; Swagger Try-it defaults to Bearer instead.
...
  responses:
    Error:
      description: Error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
      required: [error]
...
    CreatePlaylistRequest:
      type: object
      required: [seed_track_id]
      properties:
        seed_track_id:
          type: integer
          format: int64
          minimum: 1
        length:
          type: integer
        name:
          type: string
    CreateUserPlaylistRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
          minLength: 1
...
    MyPlaylistList:
      type: object
      properties:
        playlists:
          type: array
          items:
            $ref: '#/components/schemas/PlaylistHeader'
    PlaylistHeader:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        seed_track_id:
          type: integer
          format: int64
          nullable: true
        user_id:
          type: integer
          format: int64
          nullable: true
        kind:
          type: string
          enum: [neighbor, user]
        length:
          type: integer
        created_at:
          type: string
        updated_at:
          type: string
    PlaylistTrack:
      type: object
      properties:
        position:
          type: integer
        track_id:
          type: integer
          format: int64
        score:
          type: number
          nullable: true
        title:
          type: string
        artist:
          type: string
        album:
          type: string
        duration_ms:
          type: integer
          format: int64
    Playlist:
      allOf:
        - $ref: '#/components/schemas/PlaylistHeader'
        - type: object
          properties:
            tracks:
              type: array
              items:
                $ref: '#/components/schemas/PlaylistTrack'
...
security:
  - bearerAuth: []